Sounds and High Scores
A bell will now right when a target is hit.
A high score appears on the left side of the target.html. This will
display the high score for the current game that the user is playing.
Easy-accuracy is the default game, and therefore the default high score
to be displayed.